refactor(team): remove default dashboard field and methods
