+++ dashboard/components.d.ts
@@ -228,9 +228,8 @@ declare module 'vue' {
     TextInput: typeof import('./node_modules/frappe-ui/src/components/TextInput.vue')['default']
     'TextInput.story': typeof import('./node_modules/frappe-ui/src/components/TextInput.story.vue')['default']
     Toast: typeof import('./node_modules/frappe-ui/src/components/Toast.vue')['default']
-    Tooltip: typeof import('./node_modules/frappe-ui/src/components/Tooltip.vue')['default']
-    'Tooltip.story': typeof import('./node_modules/frappe-ui/src/components/Tooltip.story.vue')['default']
-    TryNewDashboardButton: typeof import('./src/components/TryNewDashboardButton.vue')['default']
+    Tooltip: typeof import('./node_modules/frappe-ui/src/components/Tooltip/Tooltip.vue')['default']
+    'Tooltip.story': typeof import('./node_modules/frappe-ui/src/components/Tooltip/Tooltip.story.vue')['default']
     Underline: typeof import('./node_modules/frappe-ui/src/components/TextEditor/icons/underline.vue')['default']
     UnionPay: typeof import('./src/components/icons/cards/UnionPay.vue')['default']
     UpdateBillingDetails: typeof import('./src/components/UpdateBillingDetails.vue')['default']
