+++ dashboard/src2/data/team.js
@@ -9,8 +9,7 @@ export function getTeam() {
 			name: getCurrentTeam(),
 			whitelistedMethods: {
 				getTeamMembers: 'get_team_members',
-				removeTeamMember: 'remove_team_member',
-				changeDefaultDashboard: 'change_default_dashboard'
+				removeTeamMember: 'remove_team_member'
 			}
 		});
 	}
