+++ press/press/doctype/team/team.json
@@ -21,7 +21,6 @@
   "is_code_server_user",
   "free_account",
   "via_erpnext",
-  "default_to_new_dashboard",
   "section_break_6",
   "team_members",
   "section_break_tdm9",
@@ -411,12 +410,6 @@
   {
    "fieldname": "column_break_wejg",
    "fieldtype": "Column Break"
-  },
-  {
-   "default": "0",
-   "fieldname": "default_to_new_dashboard",
-   "fieldtype": "Check",
-   "label": "Default to New Dashboard"
   }
  ],
  "links": [
@@ -471,7 +464,7 @@
    "link_fieldname": "team"
   }
  ],
- "modified": "2024-03-23 14:15:04.526141",
+ "modified": "2024-04-04 22:20:46.667608",
  "modified_by": "Administrator",
  "module": "Press",
  "name": "Team",
